fix(settle): AR settlement Decimal end-to-end -- invoices with cents (PPN ,75) reach paid with AR 0 (unit 6b)
Law 25: Decimal inside, float at the edge. Before this, an invoice whose total carries
cents (10.533.787,75) could not be settled by any path:
- SI record-payment: InvoicePaymentCreate.amount int -> 422 on ,75; remaining int() -> 0,75 unpayable.
- deposit apply/refund/auto-apply: helpers returned int() -> applied 10.533.787, invoice
  flagged 'paid' while compute_ar_outstanding still shows 0,75; DP remaining 466.213 not 466.212,25.
- receive-payments create-with-allocations: its own get_invoice_remaining_from_journal
  returned float (387d66e0, 14 Sep) -> "remaining_after": float - Decimal = TypeError -> 500
  for EVERY allocation, cents or not. Same float in void (invoice_total - journal_remaining).
  Latent: live FE settles through the SI shortcut; 0 receive-payment creates since 14 Sep.

Changes: three deposit helpers + receive_payments helper return Decimal; plan/apply/reverse
cores Decimal; ApplyDepositItem/Refund/InvoicePaymentCreate amounts Decimal; response
amounts float; idempotency key for whole amounts unchanged (int form); money messages
formatted id-ID with ,00 dropped.

Gate (scratch milkydb_ppngate, real handlers): ,75 and whole-rupiah invoices through
(a) deposit auto-apply, (b) receive-payment + (b2) its void, (c) SI record-payment;
AR measured by compute_ar_outstanding SQL and DP by raw journal net (not the helper under test).
base 3bbe8df1 RED (b,b2,c ,75 and a) / new GREEN 9/9 /
sab (deposit helpers int, schemas widened) RED on (a) only /
sab2 (receive_payments helper float: cents kept, amounts agree) RED on (b,b2) 500.

(cherry picked from commit 110df638bf4018bc1ed45e9264f6fa2bac180cee)

+++ backend/api_gateway/app/schemas/sales_invoices.py
@@ -85,7 +85,7 @@ class InvoiceItemResponse(BaseModel):
 class InvoicePaymentCreate(BaseModel):
     """Payment record for creation."""
 
-    amount: int = Field(..., gt=0)
+    amount: Decimal = Field(..., gt=0)  # 6b: sen boleh (faktur ber-PPN ,75)
     payment_date: date
     payment_method: Literal["cash", "transfer", "check", "other"]
     account_id: str = Field(..., description="CoA account ID for Kas/Bank")
@@ -98,7 +98,7 @@ class InvoicePaymentResponse(BaseModel):
     """Payment record response."""
 
     id: str
-    amount: int
+    amount: float
     payment_date: str
     payment_method: str
     account_id: str
